Move map blocs in pouvoirs/local page
